+++ resources/js/components/twc-ui/alert.tsx
@@ -4,13 +4,13 @@ import { cn } from '@/lib/utils'
 import { Icon, type IconType } from './icon'
 
 const alertVariants = cva(
-  'relative w-full rounded-lg border px-4 py-3 text-sm grid has-[>svg]:grid-cols-[calc(var(--spacing)*4)_1fr] grid-cols-[0_1fr] has-[>svg]:gap-x-3 gap-y-0.5 items-start [&>svg]:size-4 [&>svg]:translate-y-0.5 [&>svg]:text-current',
+  'relative w-full rounded-lg border px-4 py-3 text-sm grid has-[>svg]:grid-cols-[calc(var(--spacing)*4)_1fr] grid-cols-[0_1fr] has-[>svg]:gap-x-3 gap-y-0.5 items-start [&>svg]:size-5 [&>svg]:translate-y-0.5 [&>svg]:text-current',
   {
     variants: {
       variant: {
         default: 'bg-card text-card-foreground',
         destructive:
-          'bg-destructive/100 text-destructive bg-card [&>svg]:text-current *:data-[slot=alert-description]:text-destructive/90',
+          'bg-destructive/5 border-destructive/20 text-destructive [&>svg]:text-current *:data-[slot=alert-description]:text-destructive/90',
         info: 'bg-yellow-50'
       }
     },
@@ -33,7 +33,7 @@ const Alert: React.FC<AlertProps> = ({ className, variant, icon, children, title
       className={cn(alertVariants({ variant }), className)}
       {...props}
     >
-      {icon && <Icon icon={icon} className="size-4" />}
+      {icon && <Icon icon={icon} />}
       {title && <AlertTitle>{title}</AlertTitle>}
       <AlertDescription>{children}</AlertDescription>
     </div>
@@ -44,7 +44,10 @@ const AlertTitle: React.FC<React.ComponentProps<'div'>> = ({ className, ...props
   return (
     <div
       data-slot="alert-title"
-      className={cn('col-start-2 line-clamp-1 min-h-4 font-medium tracking-tight', className)}
+      className={cn(
+        'col-start-2 line-clamp-1 min-h-4 pt-0.5 font-medium tracking-tight',
+        className
+      )}
       {...props}
     />
   )