+++ resources/js/components/twc-ui/date-field.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import type { DateValue } from '@internationalized/date'
 import type { VariantProps } from 'class-variance-authority'
 import type React from 'react'